+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## v1.9.0 - 2022-04-27
+### Added
+- `GET /sapi/v1/managed-subaccount/accountSnapshot` to support investor master account query asset snapshot of managed sub-account.
+- `GET /sapi/v1/margin/rateLimit/order` to display the user's current margin order count usage for all intervals.
+- `GET /sapi/v1/portfolio/account` to support query portfolio margin account info.
+
+### Changed
+- `POST /sapi/v1/bswap/liquidityAdd` - New optional parameter `type`
+- `POST /sapi/v1/sub-account/universalTransfer` - New transfer types `MARGIN`,`ISOLATED_MARGIN` and parameter `symbol`
+- `GET /api/v3/depth` - The `limit` value can be outside of the previous values (i.e. 5, 10, 20, 50, 100, 500, 1000, 5000). The limit still cannot exceed 5000.
+- `POST /api/v3/order`, `POST /api/v3/order/test` and `/api/v3/order/oco` - New optional parameter `trailingDelta`  
+
 ## v1.8.0 - 2022-02-24
 ### Added
 - `POST /sapi/v1/giftcard/createCode` to create a code for Binance Gift Card